test_shared_constants_unification_not_connected_const
