+++ tests/torch/data/reference_graphs/fx/transformed/shared_constants_unification_not_connected_const.dot
@@ -1,0 +1,20 @@
+strict digraph  {
+"0 linear_weight" [id=0, type=get_attr];
+"1 linear_bias" [id=1, type=get_attr];
+"2 lm_head_weight_1" [id=2, type=get_attr];
+"3 lm_head_bias" [id=3, type=get_attr];
+"4 input_ids" [id=4, type=input];
+"5 embedding" [id=5, type=embedding];
+"6 linear" [id=6, type=linear];
+"7 linear_1" [id=7, type=linear];
+"8 output_1" [id=8, type=output];
+"0 linear_weight" -> "6 linear"  [label="(8, 8)", style=solid];
+"1 linear_bias" -> "6 linear"  [label="(8,)", style=solid];
+"2 lm_head_weight_1" -> "5 embedding"  [label="(16, 8)", style=solid];
+"2 lm_head_weight_1" -> "7 linear_1"  [label="(16, 8)", style=solid];
+"3 lm_head_bias" -> "7 linear_1"  [label="(16,)", style=solid];
+"4 input_ids" -> "5 embedding"  [label="(8,)", style=solid];
+"5 embedding" -> "6 linear"  [label="(8, 8)", style=solid];
+"6 linear" -> "7 linear_1"  [label="(8, 8)", style=solid];
+"7 linear_1" -> "8 output_1"  [label="(8, 16)", style=solid];
+}
